if thread is archived, remove components only

+++ src/main/java/mandarin/packpack/supporter/server/holder/component/ConfirmButtonHolder.java
@@ -3,6 +3,8 @@ package mandarin.packpack.supporter.server.holder.component;
 import common.CommonStatic;
 import mandarin.packpack.supporter.lang.LangID;
 import net.dv8tion.jda.api.entities.Message;
+import net.dv8tion.jda.api.entities.channel.concrete.ThreadChannel;
+import net.dv8tion.jda.api.entities.channel.unions.MessageChannelUnion;
 import net.dv8tion.jda.api.events.interaction.component.GenericComponentInteractionCreateEvent;
 import org.jetbrains.annotations.NotNull;
 
@@ -21,7 +23,14 @@ public class ConfirmButtonHolder extends ComponentHolder {
     public void onEvent(@NotNull GenericComponentInteractionCreateEvent event) {
         switch (event.getComponentId()) {
             case "confirm" -> {
-                message.delete().queue();
+                MessageChannelUnion channel = message.getChannel();
+
+                if (channel instanceof ThreadChannel tc && tc.isArchived()) {
+                    message.editMessageComponents().queue();
+                } else {
+                    message.delete().queue();
+                }
+
                 action.run();
             }
             case "cancel" -> message.delete().queue();
